Add support for newer import API and refactor some code
