+++ addons/io_hubs_addon/__init__.py
@@ -1,5 +1,11 @@
-from .io import gltf_exporter, gltf_importer
 from . import (nodes, components)
+from .io import gltf_exporter, panels
+import bpy
+if bpy.app.version < (3, 0, 0):
+    from .io import gltf_importer_compat as gltf_importer
+else:
+    from .io import gltf_importer
+
 bl_info = {
     "name": "Hubs Blender Addon",
     "author": "Mozilla Hubs",
@@ -16,6 +22,7 @@ bl_info = {
 
 
 def register():
+    panels.register()
     gltf_exporter.register()
     gltf_importer.register()
     nodes.register()
@@ -27,6 +34,7 @@ def unregister():
     nodes.unregister()
     gltf_importer.unregister()
     gltf_exporter.unregister()
+    panels.unregister()
 
 
 # called by gltf-blender-io after it has loaded
@@ -36,6 +44,9 @@ glTF2ExportUserExtension = gltf_exporter.glTF2ExportUserExtension
 glTF2_pre_export_callback = gltf_exporter.glTF2_pre_export_callback
 glTF2_post_export_callback = gltf_exporter.glTF2_post_export_callback
 
+if bpy.app.version > (3, 0, 0):
+    glTF2ImportUserExtension = gltf_importer.glTF2ImportUserExtension
+
 
 def register_panel():
-    return gltf_exporter.register_export_panel()
+    return panels.register_panels()
