+++ addons/io_hubs_addon/components/definitions/audio_target.py
@@ -4,7 +4,7 @@ from ..hubs_component import HubsComponent
 from ..utils import has_component
 from ..types import Category, PanelType, NodeType
 from bpy.types import Object
-from ...io.utils import gather_joint_property, gather_node_property
+from ...io.utils import gather_joint_property, gather_node_property, add_hubs_import_component
 
 BLANK_ID = "374e54CMHFCipSk"
 
@@ -159,3 +159,17 @@ class AudioTarget(HubsComponent):
             'minDelay': self.minDelay,
             'debug': self.debug
         }
+
+    @classmethod
+    def gather_import(cls, import_settings, blender_object, component_name, component_value):
+        blender_component = add_hubs_import_component(
+            component_name, blender_object)
+
+        for property_name, property_value in component_value.items():
+            if property_name == 'srcNode':
+                setattr(blender_component, property_name,
+                        import_settings.vnodes[property_value['index']].blender_object)
+
+            else:
+                print(f"{property_name} = {property_value}")
+                setattr(blender_component, property_name, property_value)
