+++ addons/io_hubs_addon/components/definitions/video_texture_target.py
@@ -3,7 +3,8 @@ from ..hubs_component import HubsComponent
 from ..types import Category, PanelType, NodeType
 from ..utils import has_component
 from bpy.types import Object
-from ...io.utils import gather_joint_property, gather_node_property
+from ...io.utils import gather_joint_property, gather_node_property, add_hubs_import_component
+
 
 BLANK_ID = "pXph8WBzMu9fung"
 
@@ -149,3 +150,16 @@ class VideoTextureTarget(HubsComponent):
             'srcNode': gather_joint_property(export_settings, self.srcNode, self, 'bone') if self.bone_id != BLANK_ID else gather_node_property(
                 export_settings, object, self, 'srcNode'),
         }
+
+    @classmethod
+    def gather_import(cls, import_settings, blender_object, component_name, component_value):
+        blender_component = add_hubs_import_component(
+            component_name, blender_object)
+
+        for property_name, property_value in component_value.items():
+            if property_name == 'srcNode':
+                setattr(blender_component, property_name,
+                        import_settings.vnodes[property_value['index']].blender_object)
+
+            else:
+                setattr(blender_component, property_name, property_value)
